Use DXC for SM6 and plumb DXIL it to RT shader creation
